src/requirements.txt
python-telegram-bot[job-queue]==22.1
pandas
sqlalchemy
psycopg2-binary
python-dotenv
openai
langchain
langchain-openai
annoy
requests
